+++ CHANGELOG.md
@@ -6,7 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0),
 and this project adheres to
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [0.0.22] - 2026-08-10
 
 ### Added
 
@@ -423,7 +423,9 @@ and this project adheres to
 - ✨(onboarding) add activation code logic for launch #62
 - 💄(chat) add code highlighting for LLM responses #67
 
-[unreleased]: https://github.com/suitenumerique/conversations/compare/v0.0.21...main
+[unreleased]: https://github.com/suitenumerique/conversations/compare/v0.0.22...main
+
+[0.0.22]: https://github.com/suitenumerique/conversations/compare/v0.0.22
 
 [0.0.21]: https://github.com/suitenumerique/conversations/compare/v0.0.21